Feature/sp 323

+++ src/main/java/com/bitpay/sdk/client/RefundClient.java
@@ -14,6 +14,7 @@ import com.bitpay.sdk.model.Invoice.Refund;
 import com.bitpay.sdk.util.AccessTokens;
 import com.bitpay.sdk.util.GuidGenerator;
 import com.bitpay.sdk.util.JsonMapperFactory;
+import com.bitpay.sdk.util.ParameterAdder;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -57,12 +58,14 @@ public class RefundClient {
      * @return Refund
      * @throws BitPayException BitPayException
      */
-    public Refund createRefund(Refund refund) throws BitPayException {
-        final Map<String, Object> params = createBasicParamsForCreate(refund);
+    public Refund create(Refund refund) throws BitPayException {
+        if (Objects.isNull(refund)) {
+            throw new RefundCreationException(null, "missing required parameter");
+        }
 
+        final Map<String, Object> params = createBasicParamsForCreate(refund);
         Refund result;
         JsonMapper mapper = JsonMapperFactory.create();
-
         String json;
 
         try {
@@ -84,44 +87,6 @@ public class RefundClient {
         return result;
     }
 
-    private Map<String, Object> createBasicParamsForCreate(Refund refund) throws BitPayException {
-        String guid = Objects.isNull(refund.getGuid()) ? this.guidGenerator.execute() : refund.getGuid();
-        String invoiceId = refund.getInvoice();
-        Double amount = refund.getAmount();
-        Boolean preview = refund.getPreview();
-        Boolean immediate = refund.getImmediate();
-        Boolean buyerPaysRefundFee = refund.getBuyerPaysRefundFee();
-        String reference = refund.getReference();
-
-        if (invoiceId == null && amount == null) {
-            throw new RefundCreationException(null, "Invoice ID, amount and currency are required to issue a refund.");
-        }
-
-        final Map<String, Object> params = new HashMap<>();
-        params.put("token", this.accessTokens.getAccessToken(Facade.MERCHANT));
-        if (invoiceId != null) {
-            params.put("invoiceId", invoiceId);
-        }
-        if (amount != null) {
-            params.put("amount", amount);
-        }
-        if (preview != null) {
-            params.put("preview", preview);
-        }
-        if (immediate != null) {
-            params.put("immediate", immediate);
-        }
-        if (buyerPaysRefundFee != null) {
-            params.put("buyerPaysRefundFee", buyerPaysRefundFee);
-        }
-        if (reference != null) {
-            params.put("reference", reference);
-        }
-        params.put("guid", guid);
-
-        return params;
-    }
-
     /**
      * Retrieve a previously made refund request on a BitPay invoice.
      *
@@ -131,10 +96,14 @@ public class RefundClient {
      * @throws BitPayException      BitPayException class
      */
     public Refund getById(String refundId) throws RefundQueryException, BitPayException {
+        if (Objects.isNull(refundId)) {
+            throw new RefundQueryException(null, "missing required parameter");
+        }
+
         Refund refund;
 
         final List<BasicNameValuePair> params = new ArrayList<BasicNameValuePair>();
-        params.add(new BasicNameValuePair("token", this.accessTokens.getAccessToken(Facade.MERCHANT)));
+        ParameterAdder.execute(params,"token", this.accessTokens.getAccessToken(Facade.MERCHANT));
 
         try {
             HttpResponse response = this.bitPayClient.get("refunds/" + refundId, params, true);
@@ -159,10 +128,14 @@ public class RefundClient {
      * @throws BitPayException BitPayException
      */
     public Refund getByGuid(String guid) throws BitPayException {
+        if (Objects.isNull(guid)) {
+            throw new RefundQueryException(null, "missing required parameter");
+        }
+
         Refund refund;
 
         final List<BasicNameValuePair> params = new ArrayList<BasicNameValuePair>();
-        params.add(new BasicNameValuePair("token", this.accessTokens.getAccessToken(Facade.MERCHANT)));
+        ParameterAdder.execute(params,"token", this.accessTokens.getAccessToken(Facade.MERCHANT));
 
         try {
             HttpResponse response = this.bitPayClient.get("refunds/guid/" + guid, params, true);
@@ -189,14 +162,20 @@ public class RefundClient {
      * @throws BitPayException      BitPayException class
      */
     public List<Refund> getRefundsByInvoiceId(String invoiceId) throws RefundQueryException, BitPayException {
+        if (Objects.isNull(invoiceId)) {
+            throw new RefundQueryException(null, "missing required parameter");
+        }
+
         List<Refund> refunds;
         final List<BasicNameValuePair> params = new ArrayList<BasicNameValuePair>();
-        params.add(new BasicNameValuePair("token", this.accessTokens.getAccessToken(Facade.MERCHANT)));
-        params.add(new BasicNameValuePair("invoiceId", invoiceId));
+        ParameterAdder.execute(params, "token", this.accessTokens.getAccessToken(Facade.MERCHANT));
+        ParameterAdder.execute(params, "invoiceId", invoiceId);
 
         try {
             HttpResponse response = this.bitPayClient.get("refunds/", params, true);
-            refunds = Arrays.asList(new ObjectMapper().readValue(this.bitPayClient.responseToJsonString(response), Refund[].class));
+            refunds = Arrays.asList(
+                JsonMapperFactory.create().readValue(this.bitPayClient.responseToJsonString(response), Refund[].class)
+            );
         } catch (JsonProcessingException e) {
             throw new RefundQueryException(null,
                 "failed to deserialize BitPay server response (Refund) : " + e.getMessage());
@@ -282,12 +261,14 @@ public class RefundClient {
      * @throws BitPayException         BitPayException class
      */
     public Boolean sendRefundNotification(String refundId) throws RefundCreationException, BitPayException {
+        if (Objects.isNull(refundId)) {
+            throw new RefundCreationException(null, "missing required parameter");
+        }
+
         final Map<String, String> params = new HashMap<>();
         params.put("token", this.accessTokens.getAccessToken(Facade.MERCHANT));
 
-        Refund refund;
         JsonMapper mapper = JsonMapperFactory.create();
-
         Boolean result;
         String json;
 
@@ -322,10 +303,14 @@ public class RefundClient {
      * @throws BitPayException             BitPayException class
      */
     public Refund cancel(String refundId) throws RefundCancellationException, BitPayException {
+        if (Objects.isNull(refundId)) {
+            throw new RefundCancellationException(null, "missing required parameter");
+        }
+
         Refund refund;
 
         final List<BasicNameValuePair> params = new ArrayList<BasicNameValuePair>();
-        params.add(new BasicNameValuePair("token", this.accessTokens.getAccessToken(Facade.MERCHANT)));
+        ParameterAdder.execute(params, "token", this.accessTokens.getAccessToken(Facade.MERCHANT));
 
         try {
             HttpResponse response = this.bitPayClient.delete("refunds/" + refundId, params);
@@ -350,10 +335,14 @@ public class RefundClient {
      * @since 8.7.0
      */
     public Refund cancelByGuid(String guid) throws RefundCancellationException, BitPayException {
+        if (Objects.isNull(guid)) {
+            throw new RefundCancellationException(null, "missing required parameter");
+        }
+
         Refund refund;
 
         final List<BasicNameValuePair> params = new ArrayList<BasicNameValuePair>();
-        params.add(new BasicNameValuePair("token", this.accessTokens.getAccessToken(Facade.MERCHANT)));
+        ParameterAdder.execute(params, "token", this.accessTokens.getAccessToken(Facade.MERCHANT));
 
         try {
             HttpResponse response = this.bitPayClient.delete("refunds/guid/" + guid, params);
@@ -386,5 +375,43 @@ public class RefundClient {
         }
 
         return json;
+    }
+
+    private Map<String, Object> createBasicParamsForCreate(Refund refund) throws BitPayException {
+        String guid = Objects.isNull(refund.getGuid()) ? this.guidGenerator.execute() : refund.getGuid();
+        String invoiceId = refund.getInvoice();
+        Double amount = refund.getAmount();
+        Boolean preview = refund.getPreview();
+        Boolean immediate = refund.getImmediate();
+        Boolean buyerPaysRefundFee = refund.getBuyerPaysRefundFee();
+        String reference = refund.getReference();
+
+        if (invoiceId == null && amount == null) {
+            throw new RefundCreationException(null, "Invoice ID, amount and currency are required to issue a refund.");
+        }
+
+        final Map<String, Object> params = new HashMap<>();
+        params.put("token", this.accessTokens.getAccessToken(Facade.MERCHANT));
+        if (invoiceId != null) {
+            params.put("invoiceId", invoiceId);
+        }
+        if (amount != null) {
+            params.put("amount", amount);
+        }
+        if (preview != null) {
+            params.put("preview", preview);
+        }
+        if (immediate != null) {
+            params.put("immediate", immediate);
+        }
+        if (buyerPaysRefundFee != null) {
+            params.put("buyerPaysRefundFee", buyerPaysRefundFee);
+        }
+        if (reference != null) {
+            params.put("reference", reference);
+        }
+        params.put("guid", guid);
+
+        return params;
     }
 }
